+++ app/src/main/java/org/simple/clinic/teleconsultlog/teleconsultrecord/TeleconsultRecordModule.kt
@@ -1,0 +1,17 @@
+package org.simple.clinic.teleconsultlog.teleconsultrecord
+
+import dagger.Module
+import dagger.Provides
+import org.simple.clinic.AppDatabase
+
+@Module
+class TeleconsultRecordModule {
+  @Provides
+  fun teleconsultRecordDao(appDatabase: AppDatabase) = appDatabase.teleconsultRecordDao()
+
+  @Provides
+  fun teleconsultRecordPrescribedDrugsDao(appDatabase: AppDatabase) = appDatabase.teleconsultRecordPrescribedDrugDao()
+
+  @Provides
+  fun teleconsultRecordWithPrescribedDrugDao(appDatabase: AppDatabase) = appDatabase.teleconsultRecordWithPrescribedDrugDao()
+}