+++ app/src/main/java/org/simple/clinic/teleconsultlog/teleconsultrecord/TeleconsultRequestInfo.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.teleconsultlog.teleconsultrecord
+
+import java.time.Instant
+import java.util.UUID
+
+data class TeleconsultRequestInfo(
+    val requesterId: UUID,
+
+    val facilityId: UUID,
+
+    val requestedAt: Instant
+)